Move HMAC helper to auth package
